Implement NljToHashjoin Rule
- Implemented NljToHashjoin rule under crate::rules module.
This rule requires ANALYZE, it checks each NLJ node in the plan tree and creates a HashJoin hint if the stats meet
either of these
heuristic conditions:
1. The plan_rows exceeds a given threshold
2. The actual_rows exceeds plan_rows times a ratio
This rule uses a state to record the added hints as it traverses the
tree.
Subqueries and SetOps are yet to be supported, still need to investigate how to
add hints to a subquery.
Feel free to extend this plan or add non-ANALYZE version of this plan

- Added test for both the NljToHashjoin rule and using it in the
Optimizer.

- Formatted the repo
- added helper functions in postgres2sql.rs
- modified the imdb installation sh script to specify postgres as user