Implement CardCorrectionRule
Implemented CardCorrectionRule that takes the ANALYZE plan, add hints to
replace the estimated cardinalities of join results with the actual join
cardinality of them, multiplied by a card_multiplier specified by the
user.

When parallel execution is involved, the actual_rows field of join nodes
are inaccurate and is roughly the `true card of the join / number of
workers launched`. Currently a heuristic is set up that the
card_multiplier is multiplied by num_workers whenever it finds a Gather
node in the PlanNode tree.

In the provided test, the Rows hint would always set the CE of the root
Gather node no matter what join result is specified, this is unexpected
but acceptable as it doesn't interfere with planning much.

The added test / example uses card_correction_rule_test.sql, when the
rule is applied the outer join gets converted to HashJoin and thus give
~30% speedup. The inner join could not be converted because this hint
could only apply to join results, but not scan results.